Clean test_fs_write. NFC
- Rename test (test_write.cpp is very generic and test_fs_write matches
  the name of the test)
- Convert to C
- Format code
