+++ test/fs/test_fs_write.out
@@ -1,0 +1,1 @@
+read Hello! 1234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890
