--- test/fs/test_write.out
@@ -1,1 +1,0 @@
-read Hello! 1234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890
